+++ sdk/keyvault/azure-keyvault-secrets/samples/list_operations.py
@@ -25,7 +25,7 @@ from azure.core.exceptions import HttpResponseError
 #
 # 2. List secrets from the Key Vault (list_secrets)
 #
-# 3. List secret versions from the Key Vault (list_secret_versions)
+# 3. List secret versions from the Key Vault (list_properties_of_secret_versions)
 #
 # 4. List deleted secrets from the Key Vault (list_deleted_secrets). The vault has to be soft-delete enabled to perform
 # this operation.
@@ -70,7 +70,7 @@ def run_sample():
         # You need to check all the different values your bank account password secret had previously. Lets print all
         # the versions of this secret.
         print("\n.. List versions of the secret using its name")
-        secret_versions = client.list_secret_versions(bank_secret.name)
+        secret_versions = client.list_properties_of_secret_versions(bank_secret.name)
         for secret_version in secret_versions:
             print(
                 "Bank Secret with name '{0}' has version: '{1}'.".format(